+++ rick/constants.py
@@ -1,4 +1,4 @@
-RICK_VERSION = ["0", "8", "2"]
+RICK_VERSION = ["0", "8", "3"]
 
 
 def get_version():